refactor: project and semester routes with client association checks
- Add checks to ensure only associated clients can access or modify projects.
- Update DELETE and GET handlers to validate client and additional client roles.
- Extend test cases to cover new client association scenarios.

+++ src/app/api/projects/[id]/route.ts
@@ -1,4 +1,4 @@
-import { NextRequest, NextResponse } from 'next/server'
+import { NextResponse } from 'next/server'
 import { StatusCodes } from 'http-status-codes'
 import { NotFound } from 'payload'
 import { z, ZodError } from 'zod'
@@ -7,6 +7,7 @@ import { Security } from '@/business-layer/middleware/Security'
 import { RequestWithUser } from '@/types/Requests'
 import { MediaSchema, UserSchema } from '@/types/Payload'
 import { UserRole } from '@/types/User'
+import { User } from '@/payload-types'
 
 export const UpdateProjectRequestBody = z.object({
   name: z.string().optional(),
@@ -78,11 +79,11 @@ class RouteWrapper {
     const { id } = await params
     const projectService = new ProjectService()
     try {
-      const fetchedProject = await projectService.getProjectById(id)
+      const project = await projectService.getProjectById(id)
       if (
         req.user.role !== UserRole.Admin &&
-        fetchedProject.client !== req.user.id &&
-        !fetchedProject.additionalClients?.includes(req.user.id)
+        (project.client as User).id !== req.user.id &&
+        !(project.additionalClients as User[])?.some((client) => client.id === req.user.id)
       ) {
         return NextResponse.json({ error: 'Unauthorized' }, { status: StatusCodes.UNAUTHORIZED })
       }
@@ -113,11 +114,19 @@ class RouteWrapper {
    * @param params - The parameters object containing the project ID.
    */
   @Security('jwt', ['admin', 'client'])
-  static async DELETE(_req: NextRequest, { params }: { params: Promise<{ id: string }> }) {
+  static async DELETE(req: RequestWithUser, { params }: { params: Promise<{ id: string }> }) {
     const { id } = await params
     const projectService = new ProjectService()
 
     try {
+      const project = await projectService.getProjectById(id)
+      if (
+        req.user.role !== UserRole.Admin &&
+        (project.client as User).id !== req.user.id &&
+        !(project.additionalClients as User[])?.some((client) => client.id === req.user.id)
+      ) {
+        return NextResponse.json({ error: 'Unauthorized' }, { status: StatusCodes.UNAUTHORIZED })
+      }
       await projectService.deleteProject(id)
       return new NextResponse(null, { status: StatusCodes.NO_CONTENT })
     } catch (error) {
